+++ packages/query/src/graphql/plugins/GetMetadataPlugin.ts
@@ -62,7 +62,7 @@ async function fetchFromApi(): Promise<void> {
     meta = await fetch(new URL(`meta`, indexerUrl));
     const result = await meta.json();
     Object.assign(metaCache, result);
-  } catch (e) {
+  } catch (e: any) {
     metaCache.indexerHealthy = false;
     console.warn(`Failed to fetch indexer meta, `, e.message);
   }
@@ -70,7 +70,7 @@ async function fetchFromApi(): Promise<void> {
   try {
     health = await fetch(new URL(`health`, indexerUrl));
     metaCache.indexerHealthy = !!health.ok;
-  } catch (e) {
+  } catch (e: any) {
     metaCache.indexerHealthy = false;
     console.warn(`Failed to fetch indexer health, `, e.message);
   }
@@ -181,6 +181,7 @@ function findNodePath(nodes: readonly SelectionNode[], path: string[]): FieldNod
 
     if (!newPath.length) return found;
 
+    if (!found.selectionSet) return;
     return findNodePath(found.selectionSet.selections, newPath);
   }
 }
@@ -238,15 +239,15 @@ export const GetMetadataPlugin = makeExtendSchemaPlugin((build: Build, options) 
         _metadatas(
           after: Cursor
           before: Cursor # distinct: [_mmr_distinct_enum] = null # filter: _MetadataFilter # first: Int
-          # offset: Int
-        ): # last: Int
+          # last: Int
+        ): # offset: Int
         # orderBy: [_MetadatasOrderBy!] = [PRIMARY_KEY_ASC]
         _Metadatas
       }
     `,
     resolvers: {
       Query: {
-        _metadata: async (_parentObject, args, context, info): Promise<MetaData> => {
+        _metadata: async (_parentObject, args, context, info): Promise<MetaData | undefined> => {
           const tableExists = metadataTableSearch(build);
           if (tableExists) {
             let rowCountFound = false;