+++ packages/query/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- enable ts strict mode
+
 ## [2.12.1] - 2024-06-12
 ### Changed
 - Update `@subql/utils`